\name{bayesAB-package}
\alias{bayesAB-package}
\alias{bayesAB}
\docType{package}
\title{
Fast Bayesian Methods for A/B Testing in R
}
\description{
bayesAB provides a suite of functions for setting up, diagnosing, and interpreting A/B tests using fast Bayesian Methods. Functions include helpers for choosing and setting priors, graphical and numerical summaries of test results in a Bayesian framework, as well as functions to derive credible intervals of your observed lift.
}
\details{
\tabular{ll}{
Package: \tab bayesAB\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2016-08-98\cr
License: \tab GPL (>= 2)\cr
}
bayesTest
}
\author{
Frank Portman - http://fportman.com

Maintainer: Frank Portman <frank1214@gmail.com>
}
