% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{grab}
\alias{grab}
\title{Grab the supplied posterior from a bayesTest object}
\usage{
grab(bT, posterior)
}
\arguments{
\item{bT}{a bayesTest object}

\item{posterior}{the name of a posterior in that object (string)}
}
\value{
a \code{bayesTest} object with the posterior parameter isolated
}
\description{
Grab the supplied posterior from a bayesTest object, returning another bayesTest object.
}
\seealso{
\code{\link{combine}}
}
