% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\name{installPlumber}
\alias{installPlumber}
\title{Install the dev version of \code{plumber}.}
\usage{
installPlumber()
}
\description{
Install the development version of \code{plumber} from Github in order for \code{deployBandit} to work.
}
\note{
This is to be deprecated once \code{plumber} >= 0.3 is on CRAN.
}
\examples{
\dontrun{installPlumber()}

}

