% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBeta}
\alias{plotBeta}
\title{Plot the PDF of the Beta distribution.}
\usage{
plotBeta(alpha, beta)
}
\arguments{
\item{alpha}{α parameter of the Beta distribution.}

\item{beta}{β parameter of the Beta distribution.}
}
\value{
The PDF of Beta(α, β).
}
\description{
Plot the PDF of the Beta distribution.
}
\examples{
plotBeta(1, 1)
plotBeta(2, 5)
}

