% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotInvGamma}
\alias{plotInvGamma}
\title{Plot the PDF of the Inverse Gamma distribution.}
\usage{
plotInvGamma(shape, scale, p = 0.95)
}
\arguments{
\item{shape}{shape parameter of the Inverse Gamma distribution.}

\item{scale}{scale parameter of the Inverse Gamma distribution.}

\item{p}{control x-axis limits (default is set to view 95% of the area under the density curve)}
}
\value{
The PDF of InvGamma(shape, scale).
}
\description{
Plot the PDF of the Inverse Gamma distribution.
}
\examples{
plotInvGamma(2, 4)
plotInvGamma(1, 17)
}

