% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotLogNormal}
\alias{plotLogNormal}
\title{Plot the PDF of the Log Normal distribution.}
\usage{
plotLogNormal(mu, sigma, area = 0.99)
}
\arguments{
\item{mu}{\eqn{\mu} parameter of the Log Normal distribution.}

\item{sigma}{\eqn{\sigma} parameter of the Log Normal distribution.}

\item{area}{control x-axis limits (default is set to view 99\% of the area under the density curve)}
}
\value{
The PDF of Log Normal(\eqn{\mu}, \eqn{\sigma^2}).
}
\description{
Plot the PDF of the Log Normal distribution.
}
\examples{
plotLogNormal(1, 1)
plotLogNormal(2, 5)
}

