% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotNormal}
\alias{plotNormal}
\title{Plot the PDF of the Normal distribution.}
\usage{
plotNormal(mu, s_sq)
}
\arguments{
\item{mu}{μ parameter of the Normal distribution.}

\item{s_sq}{σ² parameter of the Normal distribution.}
}
\value{
The PDF of Normal(μ, σ²).
}
\description{
Plot the PDF of the Normal distribution.
}
\examples{
plotNormal(1, 1)
plotNormal(2, 5)
}

