% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotNormalInvGamma}
\alias{plotNormalInvGamma}
\title{Plot the bivariate PDF of the Normal Inverse Gamma Distribution.}
\usage{
plotNormalInvGamma(mu, lambda, alpha, beta)
}
\arguments{
\item{mu}{\eqn{\mu} parameter of the Normal Inverse Gamma distribution.}

\item{lambda}{\eqn{\lambda} parameter of the Normal Inverse Gamma distribution.}

\item{alpha}{\eqn{\alpha} parameter of the Normal Inverse Gamma distribution.}

\item{beta}{\eqn{\beta} parameter of the Normal Inverse Gamma distribution.}
}
\value{
The PDF of NormalInverseGamma(mu, lambda, alpha, beta)
}
\description{
Plot the bivariate PDF of the Normal Inverse Gamma Distribution.
}
\note{
This is a bivariate distribution (commonly used to model mean and
      variance of the normal distribution) and returns a 2d contour
      plot instead of a typical one dimensional PDF. You may want to experiment
      with both this distribution and the \code{plotNormal} and \code{plotInvGamma}
      outputs separately before arriving at a suitable set of priors for the
      Normal and LogNormal \code{bayesTest}.
}
\examples{
plotNormalInvGamma(3, 1, 1, 1)
}
