% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotPareto}
\alias{plotPareto}
\title{Plot the PDF of the Pareto distribution.}
\usage{
plotPareto(xm, alpha)
}
\arguments{
\item{xm}{xm parameter of the Pareto distribution.}

\item{alpha}{alpha parameter of the Pareto distribution.}
}
\value{
The PDF of Pareto(xm, alpha).
}
\description{
Plot the PDF of the Pareto distribution.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
}
\examples{
plotPareto(1, 1)
plotPareto(5, 3)
\dontrun{plotPareto(5, 3) + ggtitle('I hate the default title!')}
}
