% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotPoisson}
\alias{plotPoisson}
\title{Plot the PDF of the Poisson distribution.}
\usage{
plotPoisson(lambda, area = 0.99)
}
\arguments{
\item{lambda}{\eqn{\lambda} parameter of the Poisson distribution.}

\item{area}{control x-axis limits (default is set to view 99\% of the area under the density curve)}
}
\value{
The PDF of Poisson(\eqn{\lambda}).
}
\description{
Plot the PDF of the Poisson distribution.
}
\examples{
plotPoisson(1)
plotPoisson(5)
}

