% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{rename}
\alias{rename}
\title{Rename the posterior for a bayesTest object}
\usage{
rename(bT, newName)
}
\arguments{
\item{bT}{a bayesTest object}

\item{newName}{the new name you want for the posterior param (string)}
}
\value{
a \code{bayesTest} object with the posterior parameter renamed
}
\description{
Rename the posterior param for a bayesTest object.
}
\seealso{
\code{\link{combine}}
}
