% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bayesTest}
\alias{summary.bayesTest}
\title{Summarize bayesTest objects}
\usage{
\method{summary}{bayesTest}(object, percentLift = rep(0,
  length(object$posteriors)), credInt = rep(0.9, length(object$posteriors)),
  ...)
}
\arguments{
\item{object}{an object of class "bayesTest"}

\item{percentLift}{a vector of length(x$posteriors). Each entry corresponds to the percent lift ((A - B) / B) to summarize for for
the respective posterior in x. Note this is on a 'point' scale. percentLift = 5 implies you want to test for a 5\% lift.}

\item{credInt}{a vector of length(x$posteriors). Each entry corresponds to the width of credible interval of (A - B) / B to calculate for
the respective posterior in x. Also on a 'point' scale.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary method for objects of class "bayesTest".
}

