% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotBeta}
\alias{plotBeta}
\title{Plot the PDF of the Beta distribution.}
\usage{
plotBeta(alpha, beta)
}
\arguments{
\item{alpha}{\eqn{\alpha} parameter of the Beta distribution.}

\item{beta}{\eqn{\beta} parameter of the Beta distribution.}
}
\value{
The PDF of Beta(\eqn{\alpha}, \eqn{\beta}).
}
\description{
Plot the PDF of the Beta distribution.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
}
\examples{
plotBeta(1, 1)
plotBeta(2, 5)
\dontrun{plotBeta(2, 5) + ggtitle('I hate the default title!')}
}
