% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotGamma}
\alias{plotGamma}
\title{Plot the PDF of the Gamma distribution.}
\usage{
plotGamma(shape, rate)
}
\arguments{
\item{shape}{shape (\eqn{\alpha}) parameter of the Gamma distribution.}

\item{rate}{rate (\eqn{\beta}) parameter of the Gamma distribution.}
}
\value{
The PDF of Gamma(shape, rate).
}
\description{
Plot the PDF of the Gamma distribution.
}
\details{
Note: We use the shape/rate parametrization of Gamma. See https://en.wikipedia.org/wiki/Gamma_distribution for details.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
}
\examples{
plotGamma(1, 1)
plotGamma(2, 5)
\dontrun{plotGamma(2, 5) + ggtitle('I hate the default title!')}
}
