% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotInvGamma}
\alias{plotInvGamma}
\title{Plot the PDF of the Inverse Gamma distribution.}
\usage{
plotInvGamma(shape, scale)
}
\arguments{
\item{shape}{shape parameter of the Inverse Gamma distribution.}

\item{scale}{scale parameter of the Inverse Gamma distribution.}
}
\value{
The PDF of InvGamma(shape, scale).
}
\description{
Plot the PDF of the Inverse Gamma distribution.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
      Also note that the \code{scale} parameter of the Inverse Gamma distribution is
      analogous to the \code{beta} (or rate) parameter of the regular Gamma distribution.
      The \code{beta} parameter of the \link{plotNormalInvGamma} distribution is analogous
      to the \code{scale} parameter here.
}
\examples{
plotInvGamma(2, 4)
plotInvGamma(1, 17)
\dontrun{plotInvGamma(1, 17) + ggtitle('I hate the default title!')}
}
