% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotInvGamma}
\alias{plotInvGamma}
\title{Plot the PDF of the Inverse Gamma distribution.}
\usage{
plotInvGamma(shape, scale, area = 0.99)
}
\arguments{
\item{shape}{shape parameter of the Inverse Gamma distribution.}

\item{scale}{scale parameter of the Inverse Gamma distribution.}

\item{area}{control x-axis limits (default is set to view 99\% of the area under the density curve)}
}
\value{
The PDF of InvGamma(shape, scale).
}
\description{
Plot the PDF of the Inverse Gamma distribution.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
}
\examples{
plotInvGamma(2, 4)
plotInvGamma(1, 17)
\dontrun{plotInvGamma(1, 17) + ggtitle('I hate the default title!')}
}

