% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotLogNormal}
\alias{plotLogNormal}
\title{Plot the PDF of the Log Normal distribution.}
\usage{
plotLogNormal(mu, sigma)
}
\arguments{
\item{mu}{\eqn{\mu} parameter of the Log Normal distribution.}

\item{sigma}{\eqn{\sigma} parameter of the Log Normal distribution.}
}
\value{
The PDF of Log Normal(\eqn{\mu}, \eqn{\sigma^2}).
}
\description{
Plot the PDF of the Log Normal distribution.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
}
\examples{
plotLogNormal(1, 1)
plotLogNormal(2, .5)
\dontrun{plotLogNormal(2, .5) + ggtitle('I hate the default title!')}
}
