% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotNormal}
\alias{plotNormal}
\title{Plot the PDF of the Normal distribution.}
\usage{
plotNormal(mu, s_sq)
}
\arguments{
\item{mu}{\eqn{\mu} parameter of the Normal distribution.}

\item{s_sq}{\eqn{\sigma^2} parameter of the Normal distribution.}
}
\value{
The PDF of Normal(\eqn{\mu}, \eqn{\sigma^2}).
}
\description{
Plot the PDF of the Normal distribution.
}
\examples{
plotNormal(1, 1)
plotNormal(2, 5)
}

