% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotNormal}
\alias{plotNormal}
\title{Plot the PDF of the Normal distribution.}
\usage{
plotNormal(mu, sd)
}
\arguments{
\item{mu}{\eqn{\mu} parameter of the Normal distribution.}

\item{sd}{\eqn{\sigma} parameter of the Normal distribution.}
}
\value{
The PDF of Normal(\eqn{\mu}, \eqn{\sigma^2}).
}
\description{
Plot the PDF of the Normal distribution.
}
\note{
The output can be treated like any \code{ggplot2} object and modified accordingly.
}
\examples{
plotNormal(1, 1)
plotNormal(2, 5)
\dontrun{plotNormal(2, 5) + ggtitle('I hate the default title!')}
}
