% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDists.R
\name{plotPareto}
\alias{plotPareto}
\title{Plot the PDF of the Pareto distribution.}
\usage{
plotPareto(xm, alpha, area = 0.65)
}
\arguments{
\item{xm}{xm parameter of the Pareto distribution.}

\item{alpha}{alpha parameter of the Pareto distribution.}

\item{area}{control x-axis limits (default is set to view 65\% of the area under the density curve. Be careful tweaking this for Pareto.)}
}
\value{
The PDF of Pareto(xm, alpha).
}
\description{
Plot the PDF of the Pareto distribution.
}
\examples{
plotPareto(1, 1)
plotPareto(5, 3)
}

